# docs/api.yaml
openapi: 3.1.0
info:
  title: Conex API
  version: 0.1.0
  description: OpenAPI specification of App.
servers:
  - url: https://conex.co.cr
...
paths:
...
  /api/account:
    get:
      tags:
        - Account
      operationId: getAccount
      summary: Get account details
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
              example:
                sub: 1
                email: john@doe.com
                created_at: 1777872664
                subscription:
                  status: active
                  due_date: ''
                  plan:
                    id: free
                    name: Free
                    price:
                      amount: 0
                      currency: USD
                    billing_period:
                      unit: ''
                      count: 0
                    supports_renewal: false
                    policy:
                      max_bytes_per_site: 1048576
                      max_sites: 1
                      max_subpaths_per_site: 0
        '401':
          $ref: '#/components/responses/ReauthenticationRequired'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags:
        - Account
      operationId: deleteAccount
      summary: Delete account
      description: Deletes the authenticated account.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Account deleted.
        '401':
          $ref: '#/components/responses/ReauthenticationRequired'
        default:
          $ref: '#/components/responses/Error'
  /api/account/email/change:
    patch:
      tags:
        - Account
      operationId: requestEmailChange
      summary: Request an email address change
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeEmailRequest'
            example:
              new_email: new@email.com
      responses:
        '204':
          description: Email change request accepted.
        '401':
          $ref: '#/components/responses/ReauthenticationRequired'
        default:
          $ref: '#/components/responses/Error'
  /api/account/email/change/confirm:
    patch:
      tags:
        - Account
      operationId: confirmEmailChange
      summary: Confirm an email address change
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeEmailConfirmRequest'
            example:
              otp: 123456
      responses:
        '204':
          description: Email change confirmed.
        '401':
          $ref: '#/components/responses/ReauthenticationRequired'
        default:
          $ref: '#/components/responses/Error'
  /api/plans:
    get:
      tags:
        - Plans
      operationId: listPlans
      summary: List available plans
      responses:
        '200':
          description: Available plans and policies.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plan'
              example:
                - id: free
                  name: Free
                  price:
                    amount: 0
                    currency: USD
                  billing_period:
                    unit: ''
                    count: 0
                  supports_renewal: false
                  policy:
                    max_bytes_per_site: 1048576
                    max_sites: 1
                    max_subpaths_per_site: 0
                - id: creator
                  name: Creator
                  price:
                    amount: 2499
                    currency: USD
                  billing_period:
                    unit: year
                    count: 1
                  supports_renewal: true
                  policy:
                    max_bytes_per_site: 10485760
                    max_sites: 3
                    max_subpaths_per_site: 2
                - id: pro
                  name: Pro
                  price:
                    amount: 7999
                    currency: USD
                  billing_period:
                    unit: year
                    count: 1
                  supports_renewal: true
                  policy:
                    max_bytes_per_site: 104857600
                    max_sites: 50
                    max_subpaths_per_site: 10
        default:
          $ref: '#/components/responses/Error'
  /api/plans/create/{planID}:
    post:
      tags:
        - Plans
      operationId: createPlanOrder
      summary: Create a PayPal order for a plan
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PlanID'
      responses:
        '201':
          description: PayPal order created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayPalOrder'
        '401':
          $ref: '#/components/responses/ReauthenticationRequired'
        default:
          $ref: '#/components/responses/Error'
  /api/plans/capture/{orderID}:
    post:
      tags:
        - Plans
      operationId: capturePlanOrder
      summary: Capture a PayPal order and activate the plan
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PayPalOrderID'
      responses:
        '200':
          description: PayPal order captured and payment recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapturePaymentResponse'
        '401':
          $ref: '#/components/responses/ReauthenticationRequired'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ReauthenticationRequired:
      description: The bearer access token is missing, invalid, or expired. The client should reauthenticate the user.
      headers:
        WWW-Authenticate:
          description: Bearer authentication challenge.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
...
    PlanID:
      name: planID
      in: path
      required: true
      schema:
        type: string
        enum:
          - creator
          - pro
    PayPalOrderID:
      name: orderID
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    Account:
      type: object
      additionalProperties: false
      required:
        - sub
        - email
        - created_at
        - subscription
      properties:
        sub:
          type: integer
        email:
          type: string
          format: email
        created_at:
          type: integer
          description: Account creation timestamp.
        subscription:
          $ref: '#/components/schemas/Subscription'
    Subscription:
      type: object
      additionalProperties: false
      required:
        - status
        - due_date
        - plan
      properties:
        status:
          type: string
        due_date:
          type: string
          description: Subscription due date as YYYY-MM-DD, or empty for non-renewing plans.
        plan:
          $ref: '#/components/schemas/Plan'
    Plan:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - price
        - billing_period
        - supports_renewal
        - policy
      properties:
        id:
          type: string
        name:
          type: string
          description: Localized plan display name.
        price:
          $ref: '#/components/schemas/PlanPrice'
        billing_period:
          $ref: '#/components/schemas/BillingPeriod'
        supports_renewal:
          type: boolean
        policy:
          $ref: '#/components/schemas/PlanPolicy'
    PlanPolicy:
      type: object
      additionalProperties: false
      required:
        - max_bytes_per_site
        - max_sites
        - max_subpaths_per_site
      properties:
        max_bytes_per_site:
          type: integer
          format: int64
          minimum: 0
        max_sites:
          type: integer
          format: int64
          minimum: 0
        max_subpaths_per_site:
          type: integer
          format: int64
          minimum: 0
    PlanPrice:
      type: object
      additionalProperties: false
      required:
        - amount
        - currency
      properties:
        amount:
          type: integer
        currency:
          type: string
    BillingPeriod:
      type: object
      additionalProperties: false
      required:
        - unit
        - count
      properties:
        unit:
          type: string
        count:
          type: integer
    PayPalOrder:
      type: object
      additionalProperties: true
      required:
        - id
        - links
      properties:
        id:
          type: string
        links:
          type: array
          items:
            type: object
            additionalProperties: true
    Payment:
      type: object
      additionalProperties: false
      required:
        - order_id
        - sub
        - plan_id
        - amount
        - currency
        - status
        - created_at
        - paid_at
      properties:
        order_id:
          type: string
        sub:
          type: integer
          format: int64
        plan_id:
          type: string
        amount:
          type: integer
          format: int64
        currency:
          type: string
        status:
          type: string
        created_at:
          type: integer
          format: int64
        paid_at:
          type: integer
          format: int64
    CapturePaymentResponse:
      type: object
      additionalProperties: false
      required:
        - payment
        - paypal
      properties:
        payment:
          $ref: '#/components/schemas/Payment'
        paypal:
          type: object
          additionalProperties: true
    ChangeEmailRequest:
      type: object
      additionalProperties: false
      required:
        - new_email
      properties:
        new_email:
          type: string
          format: email
    ChangeEmailConfirmRequest:
      type: object
      additionalProperties: false
      required:
        - otp
      properties:
        otp:
          type: integer
          description: Six-digit one-time password.
          minimum: 0
          maximum: 999999
...
    Error:
      type: object
      description: Generic error payload shape when the server returns JSON errors.
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - msg
          properties:
            msg:
              type: string